Initial console-based representation of world state

+++ AIProject/AIProject/CHomeObject.cpp
@@ -1,4 +1,4 @@
-#include "CHomeObject.h";
+#include "CHomeObject.h"
 #include <algorithm>
 #include <utility>
 #include <map>
@@ -105,11 +105,11 @@ vector<CPoint> CHomeObject::ComputeNodes(vector<CPoint> relPoints)
 		
 		// If x1, y1 are within the boundary, increment their count
 		if(x1 >= LEVEL_MIN_X && x1 <= LEVEL_MAX_X && y1 >= LEVEL_MIN_Y && y1 <= LEVEL_MAX_Y)
-			++relNodes[CPoint(x1, y1)];
+			++relNodes[CPoint((int)x1, (int)y1)];
 		
 		// If x2, y2 are within the boundary, increment their count
 		if(x2 >= LEVEL_MIN_X && x2 <= LEVEL_MAX_X && y2>= LEVEL_MIN_Y && y2 <= LEVEL_MAX_Y)
-			++relNodes[CPoint(x2, y2)];
+			++relNodes[CPoint((int)x2, (int)y2)];
 	}
 
 	for(map<CPoint, int>::const_iterator i = relNodes.begin(); i != relNodes.end(); i++)
